refactor(database): sharpen reference and synthesis contracts
