+++ packages/playground/src/lib/preview/virtual-imports-plugin.ts
@@ -195,12 +195,6 @@ export function synthesizeVirtualModule(specifier: PreviewModuleId): string {
   }
 }
 
-/** Named bindings declared by a preview virtual module. */
-export function virtualModuleExportNames(specifier: PreviewModuleId): readonly string[] {
-  const spec = ALIASES[specifier];
-  return spec.kind === 'reexport' ? spec.exports : [];
-}
-
 export function virtualImportsPlugin(): esbuild.Plugin {
   const aliased = new Set(Object.keys(ALIASES));
   return {
